Pass through encoding headers
At least one server has been observed to _require_ these headers, so
passing it through will allow things to work properly when the client
sends them, which in the modern world it almost certainly does

This will break the content-type detection when the returned content has
no content-type, but that was always a workaround anyway. If we see that
in the wild, we'll revisit it.