Lower-case header names
They are supposed to be case-insensitive, but the ones we create are
usually upper-case because we pull them from the HTTP_* lines in the
environment.

At least one server has expressed an issue with upper-case, but has no
problem with lower-case. That seems safest anyway (there's a mild
convention for lower-case, stronger since HTTP/2) so lets just
lower-case all header names.