Try hard to force the server to never encode anything
We observed a server that would not return data unless it received an
Accept-encoding header, so in @12fef0e I tried to fix this by passing
Accept-encoding and Content-encoding through the proxy.

Unfortunately I didn't notice that Furl automatically decodes the
response. In this case, we end up passing the Content-encoding and
Content-length for the encoded content back to the UA, with the decoded
content. The UA usually gets confused at this point.

So this reverts that, and instead sends Accept-encoding: identity. This
should result in us always getting the raw, uncompressed data back from
the server, at the expense of more data transferred and maybe more CPU
used on the sender to decompress before sending (but still probably
negligible each way).