Optionally set Authority header on outgoing requests
We've observed requests to some CloudFront-fronted sits that seem to
require a correct Authority: header in order to avoid the CAPTCHA. This
is weird because :authority is a HTTP/2 pseudo-header and we are but a
humble HTTP/1.1 client, so maybe there's some weird heuristic in play.
But, make it optionally enabled via HOPSCOTCH_SET_AUTHORITY just until
we've run it for long enough to get a feel for it.